chore(frontend): separate login/signup/profile components
